+++ ballcat-common/ballcat-common-util/src/main/java/com/hccake/ballcat/common/util/json/TypeReference.java
@@ -1,0 +1,31 @@
+package com.hccake.ballcat.common.util.json;
+
+import java.lang.reflect.ParameterizedType;
+import java.lang.reflect.Type;
+
+/**
+ * @author lingting 2021/2/25 21:46
+ */
+public abstract class TypeReference<T> implements Comparable<TypeReference<T>> {
+
+	protected final Type type;
+
+	protected TypeReference() {
+		Type superClass = getClass().getGenericSuperclass();
+		if (superClass instanceof Class<?>) {
+			throw new IllegalArgumentException(
+					"Internal error: TypeReference constructed without actual type information");
+		}
+		type = ((ParameterizedType) superClass).getActualTypeArguments()[0];
+	}
+
+	public Type getType() {
+		return type;
+	}
+
+	@Override
+	public int compareTo(TypeReference<T> o) {
+		return 0;
+	}
+
+}